requirements.txt
torch
torchvision
pillow
requests
opencv-python
mediapipe
openai
langchain
streamlit
whisper
pyttsx3
python-dotenv
rllib
unittest